requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
arrow==1.4.0
asgiref==3.11.0
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
billiard==4.2.3
black==25.11.0
blessed==1.25.0
boto3==1.41.2
botocore==1.41.2
bpyutils==0.5.8
bs4==0.0.2
CacheControl==0.14.4
cachetools==6.2.2
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorclass==2.2.2
cron_descriptor==2.0.6
cryptography==46.0.3
debugpy==1.8.17
decorator==5.2.1
deptry==0.24.0
distro==1.9.0
dj-static==0.0.6
Django==5.2.8
django-author==1.2.0
django-celery-beat==2.8.1
django-cors-headers==4.9.0
Django-jalali-date-new==1.1
django-json-widget==2.1.0
django-mass-edit==3.5.0
django-picklefield==3.3
django-redis==6.0.0
django-timezone-field==7.1
django-user-agents==0.4.0
django_celery_results==2.6.0
djangorestframework==3.16.1
docopt==0.6.2
emoji-regex==0.1.2
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
frozenlist==1.8.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hashids==1.3.1
hpack==4.1.0
hstspreload==2025.1.1
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jalali_core==1.0.0
jdatetime==5.2.0
jedi==0.19.2
jiter==0.12.0
jmespath==1.0.1
kombu==5.5.2
Markdown==3.10
markdown-it-py==4.0.0
markdown2==2.5.4
markdownify==1.2.2
mdurl==0.1.2
mistune==3.1.4
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.5
openai==2.8.1
openpyxl==3.1.5
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
psycopg==3.2.13
psycopg-pool==3.2.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydot==4.0.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
python-markdown-math==0.9
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requirements-parser==0.13.0
rfc3986==2.0.0
rsa==4.9.1
s3transfer==0.15.0
setuptools-git==1.2
six==1.17.0
smsir-python==1.0.8
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sqlparams==6.2.0
sqlparse==0.5.3
stack-data==0.6.3
static3==0.7.0
terminaltables==3.1.10
tomli==2.3.0
tornado==6.5.2
tqdm==4.67.1
types-python-dateutil==2.9.0.20251115
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
urllib3==2.5.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
zstandard==0.25.0
